+++ ios/Classes/SwiftGeolocationPlugin.swift
@@ -12,15 +12,15 @@ public class SwiftGeolocationPlugin: NSObject, FlutterPlugin, UIApplicationDeleg
   
   internal let registrar: FlutterPluginRegistrar
   private let locationClient = LocationClient()
-  private let locationChannel: LocationChannel
+  private let locationChannels: LocationChannels
   
   init(registrar: FlutterPluginRegistrar) {
     self.registrar = registrar
-    self.locationChannel = LocationChannel(locationClient: locationClient)
+    self.locationChannels = LocationChannels(locationClient: locationClient)
     super.init()
     
     registrar.addApplicationDelegate(self)
-    locationChannel.register(on: self)
+    locationChannels.register(on: self)
   }
   
   public static func register(with registrar: FlutterPluginRegistrar) {